requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via
    #   virtualenv
    #   webdrivermanager
appium-python-client==1.0.2
    # via robotframework-appiumlibrary
beautifulsoup4==4.9.3
    # via webdrivermanager
certifi==2020.6.20
    # via requests
cffi==1.14.5
    # via cryptography
chardet==3.0.4
    # via requests
coverage==5.5
    # via robotframework-jsonlibrary
cryptography==3.4.6
    # via pyjwt
decorator==4.4.2
    # via
    #   jsonpath-rw
    #   robotframework-appiumlibrary
distlib==0.3.1
    # via virtualenv
docutils==0.16
    # via robotframework-appiumlibrary
filelock==3.0.12
    # via virtualenv
idna==2.10
    # via requests
jsonpath-rw==1.4.0
    # via
    #   jsonpath-rw-ext
    #   robotframework-jsonlibrary
jsonpath-rw-ext==1.2.2
    # via robotframework-jsonlibrary
kitchen==1.2.6
    # via robotframework-appiumlibrary
lxml==4.6.2
    # via webdrivermanager
msedge-selenium-tools==3.141.3
    # via robotframework-zoomba
pbr==5.5.1
    # via jsonpath-rw-ext
pluggy==0.13.1
    # via tox
ply==3.11
    # via jsonpath-rw
psutil==5.8.0
    # via robotframework-zoomba
py==1.10.0
    # via tox
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.0.1
    # via -r requirements.in
python-dateutil==2.8.1
    # via robotframework-zoomba
requests==2.25.1
    # via
    #   robotframework-requests
    #   robotframework-zoomba
    #   webdrivermanager
robotframework==3.2.2
    # via
    #   -r requirements.in
    #   robotframework-appiumlibrary
    #   robotframework-jsonlibrary
    #   robotframework-requests
    #   robotframework-seleniumlibrary
    #   robotframework-sudslibrary-aljcalandra
    #   robotframework-zoomba
robotframework-appiumlibrary==1.5.0.7
    # via robotframework-zoomba
robotframework-jsonlibrary==0.3.1
    # via -r requirements.in
robotframework-pythonlibcore==2.2.1
    # via robotframework-seleniumlibrary
robotframework-requests==0.7.2
    # via robotframework-zoomba
robotframework-seleniumlibrary==5.0.0
    # via robotframework-zoomba
robotframework-sudslibrary-aljcalandra==1.1
    # via robotframework-zoomba
robotframework-zoomba==2.8.1
    # via -r requirements.in
selenium==3.141.0
    # via
    #   appium-python-client
    #   msedge-selenium-tools
    #   robotframework-appiumlibrary
    #   robotframework-seleniumlibrary
    #   robotframework-zoomba
six==1.15.0
    # via
    #   jsonpath-rw
    #   python-dateutil
    #   robotframework-appiumlibrary
    #   tox
    #   virtualenv
soupsieve==2.2
    # via beautifulsoup4
suds-jurko==0.6
    # via robotframework-sudslibrary-aljcalandra
tox==3.0.0
    # via robotframework-jsonlibrary
tqdm==4.57.0
    # via webdrivermanager
urllib3==1.25.10
    # via
    #   requests
    #   selenium
virtualenv==20.4.3
    # via tox
webdrivermanager==0.10.0
    # via -r requirements.in
